XGBoost MultiClass One-Hot Example
This is mostly forked from a kaggle script, but this is a very cleaned-up, commented version that adds extra variables and simplifies much of the transformation.  I also added some code for xgboost's variable significance, although it throws a warning so I'm not sure it's working. Still, the original URL is attributed inline.  It also doesn't perform great, but it's a good example of one-hot encoding across a multi-file many-to-many data structure with xgboost and a multiclass-logloss submission.  Hope this helps.